Add routing and weather API integration
