--- accesibilidad-demo/src/App.js
@@ -1,14 +1,0 @@
-import React from "react";
-import AccessibilityPanel from "./AccessibilityPanel";
-import Dashboard from "./Dashboard"; // importa el dashboard
-
-function App() {
-  return (
-    <div>
-      <AccessibilityPanel />
-      <Dashboard />
-    </div>
-  );
-}
-
-export default App;
