+++ accesibilidad-demo/src/domain/accessibilityProfiles.js
@@ -1,0 +1,13 @@
+export const accessibilityProfiles = {
+  blindness: [
+    "screen", "eye", "moon", "contrast", "readable", "link", "heading"
+  ],
+  motor: ["keyboard", "smartnav", "virtualkb", "enlarge"],
+  daltonism: ["contrast"],
+  visual: ["zoom", "eye", "readable", "fontIncrease"],
+  epilepsy: ["block"],
+  adhd: ["guide", "focus", "heading", "link"],
+  learning: ["speaker", "dict", "doc"],
+  elder: ["zoom", "readable", "fontIncrease"],
+  dyslexia: ["readable", "fontIncrease", "lineSpace"]
+};
