+++ accesibilidad-demo/src/domain/icons.js
@@ -1,0 +1,9 @@
+export const icons = {
+  blindness: "🦯", motor: "🦾", daltonism: "🌈", visual: "👓", epilepsy: "⚡",
+  adhd: "💡", learning: "📚", elder: "🧓", dyslexia: "🔤", screen: "🦻",
+  keyboard: "⌨️", smartnav: "↔️", speaker: "🔊", mic: "🎙️", eye: "👁️",
+  moon: "🌙", sun: "☀️", drop: "💧", contrast: "🟪", font: "🔤", cursor: "🖱️",
+  block: "🚫", cc: "📝", zoom: "🔍", readable: "🔠", image: "🖼️", link: "🔗",
+  heading: "🅷", enlarge: "⬆️", doc: "📄", mute: "🔇", focus: "🔲", guide: "📕",
+  dict: "📖", virtualkb: "🖥️"
+};
