+++ accesibilidad-demo/src/domain/mockWeather.js
@@ -1,0 +1,22 @@
+export const mockWeather = {
+  temperature: "76°C",
+  humidity: "65%",
+  wind: "15 km/h",
+  pressure: "1013 hPA",
+  air: {
+    co: "76.2",
+    no2: "76.2",
+    ozone: "76.2",
+    pm25: "76.2",
+  },
+  water: {
+    temp: "76°C",
+    oxygen: "8.8",
+    salinity: "36.5ppt",
+    chlorophyll: "76.2",
+  },
+  alerts: {
+    temp: "BAJA",
+    wind: "MEDIA",
+  },
+};
